Rename internal to registry fetcher
