+++ src/eunomia/fetchers/__init__.py
@@ -1,12 +1,12 @@
 from .factory import FetcherFactory
-from .internal import InternalFetcher, InternalFetcherConfig, internal_router_factory
+from .registry import RegistryFetcher, RegistryFetcherConfig, registry_router_factory
 
 # Register the built-in fetchers
 FetcherFactory.register_fetcher(
-    "internal",
-    InternalFetcher,
-    InternalFetcherConfig,
-    router_factory=internal_router_factory,
+    "registry",
+    RegistryFetcher,
+    RegistryFetcherConfig,
+    router_factory=registry_router_factory,
 )
