--- src/eunomia/fetchers/internal/__init__.py
@@ -1,4 +1,0 @@
-from .main import InternalFetcher, InternalFetcherConfig
-from .router import internal_router_factory
-
-__all__ = ["InternalFetcher", "InternalFetcherConfig", "internal_router_factory"]
