+++ src/eunomia/fetchers/registry/__init__.py
@@ -1,0 +1,4 @@
+from .main import RegistryFetcher, RegistryFetcherConfig
+from .router import registry_router_factory
+
+__all__ = ["RegistryFetcher", "RegistryFetcherConfig", "registry_router_factory"]
